python-whisper/requirements.txt
openai-whisper
torch
torchaudio
numpy
scipy
ffmpeg-python
flask
flask-cors
requests
sentence-transformers
transformers
scikit-learn
